pyproject.toml
[tool.poetry]
name = "supersaas-playground"
version = "0.1.0"
description = ""
authors = ["BerndA <BerndA@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
supersaas-api-client = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
